Reimplement preview without Jackson in embulk-core
