+++ embulk-core/src/main/java/org/embulk/EmbulkRunner.java
@@ -15,8 +15,8 @@ import org.embulk.config.ConfigException;
 import org.embulk.config.ConfigSource;
 import org.embulk.config.DataSource;
 import org.embulk.deps.config.YamlProcessor;
-import org.embulk.deps.preview.PreviewPrinter;
 import org.embulk.exec.ExecutionResult;
+import org.embulk.exec.PreviewPrinter;
 import org.embulk.exec.PreviewResult;
 import org.embulk.exec.ResumeState;
 import org.embulk.exec.TransactionStage;
